Move non-reducers to /helpers
