+++ src/code/reducers/helpers/challenge-progress.js
